% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babylon.R
\name{as_babylon_points}
\alias{as_babylon_points}
\title{Convert a 3-column matrix into a Babylonian point-cloud specification}
\usage{
as_babylon_points(x, color = "black", size = 0.02, alpha = 1)
}
\arguments{
\item{x}{A numeric matrix with three columns.}

\item{color}{Point color, or a vector of colors with one entry per row.}

\item{size}{Billboard point size relative to scene radius.}

\item{alpha}{Point opacity.}
}
\description{
Convert a 3-column matrix into a Babylonian point-cloud specification
}
