% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babylon.R
\name{babylon}
\alias{babylon}
\title{BabylonJS Widget}
\usage{
babylon(
  data = list(list(type = "sphere", diameter = 2)),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A list of data to be passed to the widget.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{elementId}{The ID of the HTML element to contain the widget.}
}
\description{
This function creates a new BabylonJS scene.
}
