% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene-core.R
\name{clear_scene3d}
\alias{clear_scene3d}
\title{Clear the current Babylonian scene accumulator}
\usage{
clear_scene3d()
}
\description{
This clears the in-memory scene state used by `plot3d(..., add = TRUE)` and
helper wrappers such as [points3d()] and [spheres3d()].
}
\seealso{
[current_scene3d()]
}
