% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babylon.R
\name{create_pose_3d}
\alias{create_pose_3d}
\title{Interactively pose a 3D scene and return its view parameters}
\usage{
create_pose_3d(x, width = NULL, height = NULL, ...)
}
\arguments{
\item{x}{A supported `plot3d()` object.}

\item{width}{Widget width.}

\item{height}{Widget height.}

\item{...}{Additional arguments passed to [plot3d()] with `add = FALSE`.}
}
\description{
This opens a Shiny gadget with a Babylonian scene, lets you rotate and zoom
the object, and returns the resulting `par3d()`-style view settings when you
finish.
}
