% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene-core.R
\name{last_par3d}
\alias{last_par3d}
\title{Get the last Babylonian scene view state}
\usage{
last_par3d(live = FALSE)
}
\arguments{
\item{live}{Whether to prefer the last live camera state reported back from a
Shiny-backed Babylonian widget. If unavailable, the last constructed scene
view is returned.}
}
\description{
Get the last Babylonian scene view state
}
