% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{light3d_spot}
\alias{light3d_spot}
\title{Create a BabylonJS spotlight}
\usage{
light3d_spot(
  position = c(0, 1, 0),
  direction = c(0, -1, 0),
  intensity = 1,
  diffuse = "white",
  specular = "white",
  angle = pi/3,
  exponent = 1,
  range = NULL,
  name = NULL,
  enabled = TRUE,
  add = TRUE,
  axes = TRUE,
  nticks = 5,
  ...
)
}
\arguments{
\item{position}{Optional light position for point, spot, and directional
lights.}

\item{direction}{Optional light direction for directional, spot, and
hemispheric lights.}

\item{intensity}{Light intensity multiplier.}

\item{diffuse}{Diffuse light color.}

\item{specular}{Specular light color.}

\item{angle}{Optional spotlight cone angle in radians.}

\item{exponent}{Optional spotlight falloff exponent.}

\item{range}{Optional light attenuation range.}

\item{name}{Optional light name.}

\item{enabled}{Whether the light should be enabled.}

\item{add}{Whether to add the object to the current Babylonian scene. Use
`add = FALSE` to start a fresh scene.}

\item{axes}{Whether to draw lightweight scene axes, ticks, labels, and a
bounding box.}

\item{nticks}{Approximate number of tick marks per axis when `axes = TRUE`.}

\item{...}{Reserved for future light options.}
}
\description{
Create a BabylonJS spotlight
}
