% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials-advanced.R
\name{material_ref3d}
\alias{material_ref3d}
\title{Create a named Babylon material reference}
\usage{
material_ref3d(name)
}
\arguments{
\item{name}{Registered material name.}
}
\description{
Create a named Babylon material reference
}
