% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babylon.R
\name{plot.babylon_mesh}
\alias{plot.babylon_mesh}
\title{Plot a Babylon mesh using rgl-like conventions}
\usage{
\method{plot}{babylon_mesh}(x, add = FALSE, ...)
}
\arguments{
\item{x}{A `babylon_mesh` object.}

\item{add}{Whether to add the mesh to an existing widget specification. If
`FALSE`, a new widget is returned.}

\item{...}{Additional graphical parameters. Recognized values include
`color`, `alpha`, `position`, `rotation`, and `scaling`.}
}
\description{
Plot a Babylon mesh using rgl-like conventions
}
