% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{plot3d.babylon_asset}
\alias{plot3d.babylon_asset}
\title{Plot a Babylon imported asset using rgl-like conventions}
\usage{
\method{plot3d}{babylon_asset}(x, add = FALSE, axes = TRUE, nticks = 5, ...)
}
\arguments{
\item{x}{A `babylon_asset` object created by [import_model3d()].}

\item{add}{Whether to add the asset to an existing widget specification. If
`FALSE`, a new widget is returned.}

\item{axes}{Whether to draw lightweight scene axes, ticks, labels, and a
bounding box.}

\item{nticks}{Approximate number of tick marks per axis when `axes = TRUE`.}

\item{...}{Additional graphical parameters. Recognized values include
`position`, `rotation`, `scaling`, `name`, `material`, and
`preserve_materials`.}
}
\description{
Plot a Babylon imported asset using rgl-like conventions
}
