% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{plot3d.babylon_mesh}
\alias{plot3d.babylon_mesh}
\title{Plot a Babylon mesh using rgl-like conventions}
\usage{
\method{plot3d}{babylon_mesh}(x, add = FALSE, axes = TRUE, nticks = 5, specularity = "black", ...)
}
\arguments{
\item{x}{A `babylon_mesh` object.}

\item{add}{Whether to add the mesh to an existing widget specification. If
`FALSE`, a new widget is returned.}

\item{axes}{Whether to draw lightweight scene axes, ticks, labels, and a
bounding box.}

\item{nticks}{Approximate number of tick marks per axis when `axes = TRUE`.}

\item{...}{Additional graphical parameters. Recognized values include
`color`, `alpha`, `specularity`, `position`, `rotation`, `scaling`,
`name`, and `wireframe`.}
}
\description{
Plot a Babylon mesh using rgl-like conventions
}
