% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babylon.R
\name{plot3d}
\alias{plot3d}
\title{Plot 3D objects with BabylonJS}
\usage{
plot3d(x, ...)
}
\arguments{
\item{x}{A supported 3D object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
This generic mirrors the feel of `rgl::plot3d()` while dispatching to
Babylonian renderers.
}
