% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene-effects.R
\name{postprocess3d}
\alias{postprocess3d}
\title{Create a scene post-process descriptor}
\usage{
postprocess3d(type, ...)
}
\arguments{
\item{type}{Post-process type. Currently `"depth_of_field"` is supported.}

\item{...}{Additional type-specific parameters.}
}
\description{
Create a scene post-process descriptor
}
