% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials-advanced.R
\name{register_material3d}
\alias{register_material3d}
\title{Register a reusable named material}
\usage{
register_material3d(name, material, overwrite = TRUE)
}
\arguments{
\item{name}{Registry key.}

\item{material}{Babylonian material descriptor.}

\item{overwrite}{Whether to overwrite an existing entry.}
}
\description{
Register a reusable named material
}
