% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials-advanced.R
\name{remove_material3d}
\alias{remove_material3d}
\title{Remove a registered named material}
\usage{
remove_material3d(name)
}
\arguments{
\item{name}{Registry key.}
}
\description{
Remove a registered named material
}
