% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-utils.R
\name{replace_geometry3d}
\alias{replace_geometry3d}
\title{Replace editable mesh geometry inside an imported asset}
\usage{
replace_geometry3d(x, geometry, target = NULL)
}
\arguments{
\item{x}{An imported asset created by [import_model3d()] or [import_mesh()].}

\item{geometry}{Replacement geometry as a `babylon_geometry` object or
`mesh3d`.}

\item{target}{Optional mesh selector. Defaults to `geometry$name` when
available.}
}
\description{
Replace editable mesh geometry inside an imported asset
}
