% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-utils.R
\name{scale_geometry3d}
\alias{scale_geometry3d}
\title{Scale extracted mesh geometry}
\usage{
scale_geometry3d(x, by)
}
\arguments{
\item{x}{A `babylon_geometry` object or `mesh3d`.}

\item{by}{Scalar or length-3 numeric scale.}
}
\description{
Scale extracted mesh geometry
}
