% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-utils.R
\name{set_vertices3d}
\alias{set_vertices3d}
\title{Replace the vertex matrix of extracted mesh geometry}
\usage{
set_vertices3d(x, vertices)
}
\arguments{
\item{x}{A `babylon_geometry` object or `mesh3d`.}

\item{vertices}{Replacement vertex matrix with three columns.}
}
\description{
Replace the vertex matrix of extracted mesh geometry
}
