% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-utils.R
\name{translate_geometry3d}
\alias{translate_geometry3d}
\title{Translate extracted mesh geometry}
\usage{
translate_geometry3d(x, by)
}
\arguments{
\item{x}{A `babylon_geometry` object or `mesh3d`.}

\item{by}{Length-3 numeric translation vector.}
}
\description{
Translate extracted mesh geometry
}
