% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene-core.R
\name{babylon}
\alias{babylon}
\title{BabylonJS Widget}
\usage{
babylon(
  data = list(list(type = "sphere", diameter = 2)),
  interaction = NULL,
  scene = NULL,
  sync_group = NULL,
  sync_camera = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A list of scene objects to be passed to the widget. Entries can
be Babylonian primitive specifications, file-backed meshes created with
[import_mesh()], or `mesh3d` objects such as those returned by Morpho.}

\item{interaction}{Optional interaction settings used by bespoke tools such
as landmark digitizing.}

\item{scene}{Optional scene decorations and display settings.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{elementId}{The ID of the HTML element to contain the widget.}
}
\description{
This function creates a new BabylonJS scene.
}
