% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{plot3d}
\alias{plot3d}
\title{Plot 3D objects with BabylonJS}
\usage{
plot3d(x, add = FALSE, ...)
}
\arguments{
\item{x}{A supported 3D object.}

\item{add}{Whether to add the object to the current Babylonian scene. Use
`add = FALSE` to start a fresh scene.}

\item{...}{Additional arguments passed to methods.}
}
\description{
This generic mirrors the feel of `rgl::plot3d()` while dispatching to
Babylonian renderers.
}
